extract book page rendering
